apps/ai-logic/requirements.txt
fastapi
uvicorn
llama-cloud-services
python-dotenv
llama-index-llms-openai
llama-index-core